Recommitting my changes to clabel to allow for manual labeling after fixing problems with indexing identified by Eric Firing on 19 July 2008.
In addition, I have made the following changes:

1) Placed more comments in contour.py with the intention of eventually doing
a rewrite or restructuring of that code.

2) Added two pylab_examples: contour_label_demo.py that tests out some of the
more advanced things that can now be done with contour labeling, and
ginput_manual_clabel.py that demonstrates some uses of ginput,
waitforbuttonpress, and clabel(...,manual=True).  The first of these has been
integrated into backend_driver.py, but the second cannot be because it
requires interaction.


git-svn-id: file:///mnt/workdisk/tmp/matplotlib-svn-rsync/trunk/matplotlib@5799 f61c4167-ca0d-0410-bb4a-bb21726e55ed